+++ backend/src/main/java/io/github/Lordrap2002/inventory/api/dto/ProductResponseDTO.java
@@ -1,13 +1,15 @@
 package io.github.Lordrap2002.inventory.api.dto;
 
+import java.math.BigDecimal;
+
 public class ProductResponseDTO {
     private Long id;
     private String name;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     private String description;
     private Integer stock;
 
-    public ProductResponseDTO(Long id, String name, Double unitPrice, String description, Integer stock) {
+    public ProductResponseDTO(Long id, String name, BigDecimal unitPrice, String description, Integer stock) {
         this.id = id;
         this.name = name;
         this.unitPrice = unitPrice;
@@ -23,7 +25,7 @@ public class ProductResponseDTO {
         return name;
     }
 
-    public Double getUnitPrice() {
+    public BigDecimal getUnitPrice() {
         return unitPrice;
     }
